[shill] Enable Profile objects to manage a list of Services to persist.
The Manager will hang on to a list of active services.  They will be sorted,
someday, in an order that makes sense.  Every service will be associated with
a Profile, though it may the an ephemeral profile that won't persist state to
disk.  Profiles will maintain a map of service name to service references to
track the services whose state they persist to disk.  Services may move between
Profiles, and will certainly need to be bound to one after they are registered
with the Manager, so support for this is provided as well.

BUG=chromium-os:17436
TEST=unit tests

Change-Id: Id43a0e1d97302b6f574bd2213d4f3d176bb5223f
Reviewed-on: http://gerrit.chromium.org/gerrit/4033
Reviewed-by: Chris Masone <cmasone@chromium.org>
Tested-by: Chris Masone <cmasone@chromium.org>