shill: WiFi device unregisters itself from the power manager
This bug was found in VM testing where the same WiFi device is brought up and
down multiple times.  When the WiFi device is started, it registers with the
power manager, but when it was stopped, it was no unregistering.  When it
started a second time, it tried to add itself again to the power manager, and
the power manager complained that it was already registered.

BUG=None
TEST=All unit tests.  Added a new expectation in WiFiMainTest to catch this
problem.

Change-Id: I52a3df7aaf66a702cf0ab5fed38057e9675c7fc6
Reviewed-on: https://gerrit.chromium.org/gerrit/17112
Reviewed-by: Thieu Le <thieule@chromium.org>
Commit-Ready: Gary Morain <gmorain@chromium.org>
Tested-by: Gary Morain <gmorain@chromium.org>