Get shill ready for slotted package builds.
This is the subset of https://gerrit.chromium.org/gerrit/16156
that will allow shill to build once the slotting mechanism of
https://gerrit.chromium.org/gerrit/15415 is in place. It fixes
some warnings that are the result of using slotting at all,
rather than any particular version of libchrome. This will
decouple the submission of the slotting CL from the changes
to upgrade shill to the new callback mechanism.

BUG=chromium-os:16623
TEST=Build shill and unit tests.

Change-Id: I65b52941c3a1430f0d7f6afd4033fcf1f7a2e95f
Reviewed-on: https://gerrit.chromium.org/gerrit/17253
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Eric Shienbrood <ers@chromium.org>
Commit-Ready: Eric Shienbrood <ers@chromium.org>