Don't try to create cellular device if modem interface is blacklisted.
Shill does honor the presence of usb0 on the blacklist at the time the
RTNL message is received indicating the presence of the interface.
However, there is another path that can lead to creation of a Cellular
device: when ModemManager is detected, shill calls EnumerateDevices to
get a list of modems known to ModemManager. In response, shill goes
ahead and creates a Cellular device for each modem, without checking
the blacklist.

BUG=chromium-os:28215
TEST=manual testing, with and without usb0 on the blacklist

Change-Id: I7ef4348610027d6c815b3c4d382f646666725930
Reviewed-on: https://gerrit.chromium.org/gerrit/18771
Reviewed-by: Darin Petkov <petkov@chromium.org>
Commit-Ready: Eric Shienbrood <ers@chromium.org>
Tested-by: Eric Shienbrood <ers@chromium.org>