3G: Add ability to ignore an error that occurs during asynchronous calls.
Shill performs several tasks when stopping a modem.  The first task is
to disconnect.  This task fails if the modem is not currently connected.
However, we still want to process the remaining tasks.  This CL adds the
ability to ignore errors at any particular task to allow execution of
susbsequent tasks.  Existing code that ignored errors in the callback
has been migrated to use this new mechanism.  This CL also moves the
StopModem() implementation into the base class CellularCapability.

BUG=chromium-os:28862
TEST=Unit tests, manually enable/disable cellular on device

Change-Id: I772a732c1fbcea9042328fcb10884ee2e9706a46
Reviewed-on: https://gerrit.chromium.org/gerrit/19781
Reviewed-by: Gary Morain <gmorain@chromium.org>
Commit-Ready: Thieu Le <thieule@chromium.org>
Reviewed-by: Thieu Le <thieule@chromium.org>
Tested-by: Thieu Le <thieule@chromium.org>