shill: Manager: Register services which have no profile
If we are asked over DBus to set the profile for a service and
this service is not registered, we must register the service
first.  This scenario happens when Chrome does a GetService of
a hidden network, then immediately does a SetProfile on this
Service.

BUG=chromium-os:31638
TEST=Connect to hidden network from login page in Chrome

Change-Id: If44b7800f46e61a48c398e5eeb1a0526be0adca7
Reviewed-on: https://gerrit.chromium.org/gerrit/21028
Commit-Ready: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>