shill:  Give Ethernet services an affinity to the default profile
When an ethernet service is created, save it to the default profile
by default.  This way, if static IP parameters are assigned to it,
it is shared globally instead of for only the logged-in.  This is
a reasonable default for services that are enabled by a physical
connection.  RPC callers are free to reassign the ethernet service
to a different profile, and this selection will be honored going
forwards.

BUG=chromium-os:23930
TEST=New unit tests; manual -- plug in ethernet while logged in and
ensure ethernet is added to the global profile.

Change-Id: I24e71bacc8c4bd015c2cdbb0039eb54d85dc5b65
Reviewed-on: https://gerrit.chromium.org/gerrit/21524
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>