cellular: export the DBus.ServiceProperty
Export the DBus.ServiceProperty for cellular devices, which is the
service name of the modem manager that is handling the modem.  This
allows Chrome to monitor for signals on a service name instead of a
connection name, and is consistent with the rest of Chrome's
monitoring infrastructure.

BUG=chromium-os:31757
TEST=run /usr/local/lib/flimflam/test/fake-gsm-modem --shill & list-devices
CQ-DEPEND=I0c42d65a7033beb8932303c62eb5dbfde411ea55

Change-Id: I872d9a5647fd9990f2b979e42837c46951dbd452
Reviewed-on: https://gerrit.chromium.org/gerrit/24613
Tested-by: Jason Glasgow <jglasgow@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Ready: Jason Glasgow <jglasgow@chromium.org>