shill: device_info: Postpone cdc_ether device discovery
We test cdc_ether devices at boot to decide whether they
are cellular or plain-Ethernet devices.  The method used
to do so is to figure out if there is a tty device bound
to the same USB host as the cdc_ethernet devices.  With
some systems, however, the tty devices are discovered some
time after the cdc_ether device.  To acommodate this,
postpone calling CreateDevice on such entries to allow
time for device discovery in the kernel to complete.

BUG=chromium-os:31504
TEST=Unit tests; test on physical device.

Change-Id: Ic0e18db6463519f87dde199ee7b804b2010992c6
Reviewed-on: https://gerrit.chromium.org/gerrit/26866
Commit-Ready: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>