shill: Device: Add receive / transmit byte count properties
Create "RecieveByteCount" and "TransmitByteCount" properties for
Devices, which increment persistently across shill instances.  Also
create a "ResetByteCount" method on Devices to set these counters
back to zero.

BUG=chromium-os:31584
TEST=New unit tests; list-devices on a real machine.  Restart shill,
and ensure byte counts are persisted to disk.  Write a "reset-counters"
test script and ensure that this resets the counters and the on-disk
values are set to 0 as well.

Change-Id: I8d285310d153e1e219ef523528b575e2c600de01
Reviewed-on: https://gerrit.chromium.org/gerrit/27300
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>