shill: Allow 'None' method for bgscan
Allow a special background scan method 'None' which disables background
scans entirely.  This is useful for testing behavior surrounding
background scans by comparison with runs for which we know there cannot
have been a bgscan.  It's also useful in situations where the user
explicitly doesn't want background scan to be enabled (for example in
latency sensitive applications).  We must have had something
to this effect in the past, because the autotests expect this as a possible
background scan method already.

BUG=chromium-os:32233
TEST=Run autotest 090_BgScanAbort, verify None accepted as method

Change-Id: I9fc75c541f72cf72d77cd43bae6994a10d868a0d
Reviewed-on: https://gerrit.chromium.org/gerrit/27212
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>