shill: WiFi: Remove IsConnecting
Revert to the superclass version of IsConnecting(), now that
pending services are marked in the Associating state immediately
in WiFi::ConnectTo().  To get unit tests to run, we also need
to set the service state on "surprise services" that supplicant
reports that it is roaming to.

BUG=chromium-os:23206
TEST=Run unit tests

Change-Id: Iba3c553ab5d2269e22c30c8c1e5714a24b82d2b1
Reviewed-on: https://gerrit.chromium.org/gerrit/27472
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>