shill: Service: Update RPC "Error" property on failure
When the service's failure_ enum is set, translate that into
the a string and set the DBus "Error" property so that Chrome
can display the error type as a service fails.

In order to do this, instead of using arbitrary strings, use
the string constants in service_constants.h, so that Chrome
can convert these to internationalized strings.

As part of this, realign a few of the items in the
ConnectFailure enum to match service_constants, while keeping
the same initial order so we don't mess up metrics.

BUG=chromium-os:23206
TEST=Changed unit tests, manual: connect to a non-existent hidden
network and observe "Out of range" as the error using dbus-monitor.
Also repro "Bad Passphrase" and "Out Of Range" scenarios and make
sure error mole displays the correct failure string.

Change-Id: I93f2116a101bd71a92c58ced0c80895a2091be93
Reviewed-on: https://gerrit.chromium.org/gerrit/27499
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>