shill: Unit test for 'None' bgscan method
A simple test to verify that the None background scan method is being
handled correctly.  Since this is not a real method, we're handling it
by simply refusing to append background scans when requested and None is
the method.  THis test just asserts that this is actually whats
happening.

BUG=chromium-os:32233
TEST=Ran unit tests for shill

Change-Id: Iedf1feb12fbd591a50960130b89a686a078052fb
Reviewed-on: https://gerrit.chromium.org/gerrit/27537
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Christopher Wiley <wiley@chromium.org>