shill: WiFi: Start a timer when wpa_supplicant disconnects
Currently we depend on wpa_supplicant to quickly reconnect
in situations where a disconnect occurs.  Sometimes, this
is not the case, and the time between wpa_supplicant entering
the "disconnected" state and it clearing CurrentBSS can be
a very long time indeed.  We should track any condition
where a connected service is disconnected with a timer and
manually disconnect any service that remains in this condition.

BUG=chromium-os:23207,chromium-os:32922
TEST=New unit tests; manual test in exactly the same scenario
as 32922 (actually used the same router and reproed the issue,
except that with this change we recover in 10 seconds).

Change-Id: If00da1655a972910fa0a0f9685eef0c6a96e7cdf
Reviewed-on: https://gerrit.chromium.org/gerrit/28733
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>