shill: Send Modem 'Enable' command when modem is disabled
ModemManager rejects the 'Enable' command if the modem is not
in the disabled state.  When shill attempts to enable a modem,
it first checks the modem's state.  If it is disabled, it sends
the 'Enable' command immediately.  Otherwise, it buffers the
'Enable' command and sends it when the modem transitions to
the disabled state.

This affects modemmanager-next (MM1) only.

BUG=chromium-os:23789
TEST=Added new unit test.  With Y3300, Y3400 and E362 and a locked SIM,
verify that unlocking the SIM results in the modem being successfully
enabled.

Change-Id: I922301ab7d17f44bd17ae122e80cd2889ec5790a
Reviewed-on: https://gerrit.chromium.org/gerrit/31040
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Ready: Gary Morain <gmorain@chromium.org>
Tested-by: Gary Morain <gmorain@chromium.org>