shill: Manager: Silence spammy "entry x is not registered"
These messages appear when Chrome is iterating profile entries
and we are deciding whether to furnish the properties of the
entry from a live service or from the profile contents.  It
is not a notable error when the manager cannot locate such a
Service and for systems with a large number of configured
services, this can represent a fairly large logging contingent.

BUG=None
TEST=Unit tests

Change-Id: I2b792c4bcc6ca681dc6a64aa2ea9a9c8b22e1074
Reviewed-on: https://gerrit.chromium.org/gerrit/33305
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>