shill: Modem may fail to start properly
CellularCapabilityUniversal::StartModem() only enables the modem if it
is currently disabled. If the modem is already in the registered state,
StartModem() simply completes the request and StartModemCallback() sets
the state to enabled, ignoring the actual registration state and
information.

This CL adds code that, if the modem state doesn't show up as disabled
and the underlying device state appears to be enabled, calls
GetProperties() to notify shill of any state changes that need to be
sync'd up with.

BUG=chromium-os:34028
TEST=network_3GModemControl passes with autoconnect=False option on a Y3400
modem. It still fails when autoconnect=True, which is being addressed by
chromium-os:34182

Change-Id: Ice01f6cb1ff0755416d1f046d37c1a08f15af35f
Reviewed-on: https://gerrit.chromium.org/gerrit/34064
Reviewed-by: Thieu Le <thieule@chromium.org>
Commit-Ready: Arman Uguray <armansito@chromium.org>
Tested-by: Arman Uguray <armansito@chromium.org>