shill: cellular: Suppress auto-connect when device is not running.
When a cellular device is being disabled, it may disconnect a
connected cellular service. However, shill may try to auto-connect the
service before the service is destroyed. To avoid such a race, this CL
makes a cellular service not auto-connectable when the cellular device
is not running.

BUG=chromium-os:34182
TEST=Tested the following:
1. Build and run unit tests.
2. After connecting to a cellular network, disable the cellular
   technology and verify that shill does not try to auto-connect the
   cellular service while the cellular device is being disabled.

Change-Id: I4ce18dfc416984c607976a183100779b645ebbcb
Reviewed-on: https://gerrit.chromium.org/gerrit/34497
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>
Commit-Ready: Ben Chan <benchan@chromium.org>