shill: Properly return and emit WatchedServices property
Instead of copying the "Services" property, properly return a list
of services that are in a non-"Idle/Failure/Unknown" state.  Also
emit this as a PropertyChange signal in Manager::SortServices.
Also fix a long-standing bug where EnumerateAvailableServices returned
invisble servces.

While we're here, add unit tests that assert that when a Service
state is changed, this eventually causes the above property changes
to be emitted.

BUG=chromium-os:35700,chromium-os:35701
TEST=Unit tests

Change-Id: Ifc3487f1e4856372d88da428e10d0fb40f7efd3e
Reviewed-on: https://gerrit.chromium.org/gerrit/36505
Reviewed-by: Gaurav Shah <gauravsh@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>