shill: Add dbus configuration file from shill.
The file is not installed yet -- this will be followed by an ebuild
patch. Remove "cashew" user. The new file is named after the dbus
interface and we temporarily install duplicate configs (from flimflam
and shill) but that seems to work OK.

BUG=chromium-os:35717
TEST=gmerged shill, rebooted, tested connectivity, inspected logs for
crashes or dbus issues

Change-Id: Id3b4e9882fe6d62eb963b408a1e0b7cd9865e72d
Reviewed-on: https://gerrit.chromium.org/gerrit/36772
Tested-by: Darin Petkov <petkov@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Ready: Darin Petkov <petkov@chromium.org>