shill: Disconnect before activation correctly
Allow cellular capabilities to correctly disconnect before activation by
only responding to the final disconnect notification.  Previously, shill
attempted to activate the modem upon receiving the first notification
after a disconnect operation, which is usually just a notification that
a disconnect is in progress.  Seeing that the modem hadn't actually
disconnected, shill would report the activation as failed.

The meat of this change is that CellularCapabilityCDMA stores away the
activation callback when it finds it is connected during an activation
attempt.  When the cellular device cleans up its disconnect process, it
calls down to the capability, giving us the opportunity to respond by
triggering our activation.

TEST=Unit tests, activations making use of this logic proceed correctly.
BUG=chromium-os:36301

Change-Id: I0169328403a805a06309a41b31e00b66657b9f5b
Reviewed-on: https://gerrit.chromium.org/gerrit/39028
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>
Reviewed-by: Thieu Le <thieule@chromium.org>
Commit-Ready: Christopher Wiley <wiley@chromium.org>