shill: DHCPCDProxy: Make NoReply error non-fatal in Release()
The dhcpcd process can race with us to exit when an interface goes
down or at exit.  It should not be an error if we cannot reach the
daemon to release a lease if it no longer exists.  In the worst case
we will just end up holding the lease and attempt to renew it on
re-connection, which is exactly what we'd do if we crashed.

This covers a second case (the previous being ServiceUnknown).
The NoReply case is an even smaller race, but can happen in
multiple repeated tests.

BUG=chromium-os:38887
TEST=Re-run unit tests.  Really the only good test is to watch
the crash logs and the occurrence of failures in automated tests.

Change-Id: Ic59b88245f19a63e0d8a4dc20db2d82b969666e5
Reviewed-on: https://gerrit.chromium.org/gerrit/43361
Reviewed-by: Darin Petkov <petkov@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>