shill: Refresh SIM properties when the SIM lock status changes.
On Y3300/Y3400 modems, the presence of a locked SIM prevents
CellularCapabilityUniversal from obtaining carrier related information
from the SIM (e.g. IMSI). After unlocking the SIM, shill needs to query
the SIM to obtain the latest values of the SIM DBus properties. This
data is necessary to be able to provide the correct APN to the cellular
network.

This CL addresses this issue by querying the exposed SIM DBus object to
obtain the properties in
CellularCapabilityUniversal::OnSimLockStatusChanged. This method can get
called for multiple reasons (such as the sim getting locked/unlocked,
enabling/disabling SIM locking or failed unlock attempts). Shill now
assumes that if the sim is not locked at the time of this call, it is a
good time to refresh.

BUG=chromium-os:34035
TEST=Unit test;
     1. Lock the SIM.
     2. Suspend-resume.
     3. Unlock the SIM through the UI by entering the pin.
     4. Connect to network. Connection should be successful.

     Also verify that we can successfully connect to the network with a
     non-lock-enabled sim.

Change-Id: Iebbe7548197a58a2473d6333b8321da889c5f1aa
Reviewed-on: https://gerrit.chromium.org/gerrit/40002
Commit-Queue: Arman Uguray <armansito@chromium.org>
Reviewed-by: Arman Uguray <armansito@chromium.org>
Tested-by: Arman Uguray <armansito@chromium.org>