shill: increase logging for DHCPConfig
Log more detail about when and why we try to kill the dhcpcd process.
The messages are logged at INFO (for first kill attempt), and WARNING
(for subsequent attempts).

This should not spam the log file, as stopping DHCPConfig should be an
infrequent occurence.

BUG=chromium-os:37730
TEST=unit tests

Change-Id: I0310a03d6ba8ef9c648a16590f003a851eb37122
Reviewed-on: https://gerrit.chromium.org/gerrit/41337
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>