shill: WiFiService: Load service using properties
Instead of looking for a specific storage identifier, search
for storage groups that contain the desired properties for
WiFiService::Load() and WiFiService::IsLoadableFrom().

In order to achieve this, WiFiServices should set the "SecurityClass"
property to a generalized value so they can be loaded later in
mixed wpa/rsn configurations.

BUG=chromium-os:38048
TEST=Unit Tests + Manual: Make sure PSK network stored with a
previous version of shill still auto-connects, and the
WiFi.SecurityClass property is set for the service in the
profile.

Change-Id: I95d934d829161999e6db6388e0eaeaaf3bbc06f4
Reviewed-on: https://gerrit.chromium.org/gerrit/41803
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>