shill: openvpn: Allow automatic reconnect when underlying connection reconnects.
When the underlying connection drops, restart the openvpn client, thus
resetting the connection and initiating a reconnect attempt. The
reconnect will be held until a new carrier connection is established,
up to the connect timeout of one minute.

Also, explicitly process SUCCESS management API messages so that
they're not logged as WARNINGs.

BUG=chromium-os:32416
TEST=unit tests; while connected to corp OpenVPN, switch between APs
and observe VPN reconnects; while connected to corp OpenVPN,
disconnect or disable WiFi, reconnect or re-enable it in a couple of
seconds and observe VPN reconnects.

Change-Id: I672db885d589fa020419d0badb480aee9bcc851a
Reviewed-on: https://gerrit.chromium.org/gerrit/42616
Tested-by: Darin Petkov <petkov@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Darin Petkov <petkov@chromium.org>