shill: cellular: call SetPowerState after disabling modem.
Since ModemManager no longer powers down the modem when the modem is
disabled, CellularCapabilityUniversal now calls the new SetPowerState
method on org.freedesktop.ModemManager1.Modem to put the modem into a
low power state after disabling it.

In the event that the call to SetPowerState fails, the error is ignored
and the overall disable operation is treated as success.

BUG=chromium-os:38864
TEST=1. Build and run unit tests.
     2. Enable and disable cellular through the UI. On a modem that
        supports the explicit switch to a low-power state,
        ModemManager's PowerState property should be set to low after a
        successful disable.

Change-Id: I631d8df118ed6e300ec0b197ec87e744bd2502bd
Reviewed-on: https://gerrit.chromium.org/gerrit/43241
Reviewed-by: Thieu Le <thieule@chromium.org>
Commit-Queue: Arman Uguray <armansito@chromium.org>
Tested-by: Arman Uguray <armansito@chromium.org>