shill: Configures netlink broadcast messaging with raw sockets.
This CL provides an interface for software to request the netlink
message_type and available multicast groups corresponding to a netlink
family name.  The new software completely replaces the old event_type
stuff that utilized libnl to get this information (and, as such, it is
the last nail in the coffin of using libnl for netlink communication
with the kernel).  |NetlinkMessage|s now get their message_type from
Config80211.

Also includes code to handle multi-part responses to a netlink message.

BUG=chromium:218211
TEST=unittests and manual tests.  The manual tests consist of:
  - Start shill with log=-10/wifi.  On the target, in a shell window,
    do the following:
      o stop shill
      o shill --log-level=-10 --log-scopes=wifi
  - Wait five seconds (the code will do everything it needs to at
    startup).
  - Look in /var/log/net.log and verify the following:
      o a CTRL_CMD_GETFAMILY message is sent to the kernel (just search
        for the 'CTRL_CMD_GETFAMILY' string.
      o a CTRL_CMD_NEWFAMILY message is received from the kernel
      o after those messages, verify that one or more
        NL80211_CMD_TRIGGER_SCAN messages received from the kernel.

Change-Id: I04571bdff9908ad8bd39a3a64a24e7e11074bf18
Reviewed-on: https://gerrit.chromium.org/gerrit/44770
Commit-Queue: Wade Guthrie <wdg@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>