shill: Simplifies netlink message callbacks.
Config80211 had required broadcast callbacks to be child classes of some
special (and, as it turns out, completely unnecessary) superclass.  This
CL a) replaces the superclass with |base::Callback|, b) updates
callback_metrics to use this new mechanism, and c) removes callback_object
(which was really just a template for future callback and is no longer
necessary).

BUG=None.
TEST=unittests.

Change-Id: I8f080521ba17e45e678289d10060f8e8927c0496
Reviewed-on: https://gerrit.chromium.org/gerrit/44779
Commit-Queue: Wade Guthrie <wdg@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>