shill: WiFiService: Connect to generic form of service
When connecting to a WPA/RSN/PSK service, use the generic form
when instructing wpa_supplicant to connect, so that it will connect
to either "RSN" or "WPA" networks.  There was originally a thought
to restrict this to only the network type specified via the
|security_| field in the service, but since the |security_| field
of WiFi services have historically been created and saved based on
the security mode of the first observed endpoint, this can overly
constrain which APs wpa_supplicant should be allowed to connect to.

While doing this, make sure that connection metrics refer to the
security of the AP that we've connected to instead of the (possibly
generic) version of the security of the service.

BUG=chromium-os:39813
TEST=Unit tests, network_WiFiSecMat

Change-Id: I6c192a247f10f03f88d1ec04856fcdc7bf8e96b6
Reviewed-on: https://gerrit.chromium.org/gerrit/45122
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Ashok Nagarajan <asnagarajan@chromium.org>
Tested-by: Ashok Nagarajan <asnagarajan@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>