cellular: Add support for detecting connect/disconnect loop.
This is a workaround for the E362 modem and Verizon network.  When the
SIM runs out of credits, the modem may enter a connect/disconnect loop
with the Verizon network.  This CL detects this condition and prevents
the modem from indefinitely trying to auto-connect.

BUG=chrome-os-partner:18137
TEST=Unit tests, manually test code on a device that exhibits
connect/disconnect loop

Change-Id: I61ba52194f24fbe8199da7ed6f8cc746e36c5e9c
Reviewed-on: https://gerrit.chromium.org/gerrit/45162
Reviewed-by: Arman Uguray <armansito@chromium.org>
Tested-by: Thieu Le <thieule@chromium.org>
Commit-Queue: Thieu Le <thieule@chromium.org>