shill: Support raw PEM data for CA Certificates
For all connection types that support CA Certificates (L2TP-IPSec,
OpenVPN, 802.1x WiFi) support a Service property that contains the
raw contents of the PEM certificate to be used in authenticating
the remote entity.

BUG=chromium-os:39685
TEST=Unit tests; Manual: Set the EAP.CACertPEM property on a
service and ensure that the data is stored and retrieved from
the profile correctly with newlines intact.

Change-Id: I4adc850dbb38a8b9afb55fc40260d67bcaa33485
Reviewed-on: https://gerrit.chromium.org/gerrit/45642
Commit-Queue: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>