shill: WiFiEndpoint: Update Security and NetworkMode
Retrieve new values for network_mode_ and security_mode_ in
the course of WiFiEndpoint::PropertiesChanged().  To do this,
maintain state that tracks the various hints we have retrieved
from the BSS properties we've seen so far, so we can merge
the newly retrieved data with this history to maintain an
accurate view of the current implied security mode.

BUG=chromium:206299
TEST=Unit tests, autotest (network_WiFiSecMat/120ChangeBSSSecurity)

Change-Id: Ib6a269b4cb51e0bc016335e5212b0cd21a26b404
Reviewed-on: https://gerrit.chromium.org/gerrit/46531
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>