shill: WiFiProvider: Track Endpoint -> Service mapping
Create a map for tracking the relationship between endpoints and
their associated service.  This not only allows a faster lookup,
but it also allows us to flag situations where an endpoint has
changed in a way that now violates this relationship.  In these
situations, we must now find a new service to map the endpoint to.

BUG=chromium:206299
TEST=Unit tests, autotest (network_WiFiSecMat/120ChangeBSSSecurity)

Change-Id: I7cda9d08c807acb878f7fa98c8cdaa255ba3c7f6
Reviewed-on: https://gerrit.chromium.org/gerrit/46530
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>