shill: Service: IsDependentOn(unconnected_service) should be false
A connected service should not be listed as dependent on an
un-connected service.  This previously erroneously returned
true because the equality test:

  connection_->GetLowerConnection() == b->connection()

returned true for services that had no lower connection.  This did
not matter before when the only valid sorting of services considered
their connection state first, but this is no longer the case.

BUG=chromium:224411
TEST=Unit test

Change-Id: Idcb437ba26811eac6e7476414936fea87625cb26
Reviewed-on: https://gerrit.chromium.org/gerrit/46671
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>