shill: Add ConnectionHealthChecker class to check connectivity status.
ConnectionHealthChecker asynchronously creates a TCP connection with a
remote server and checks that either (1) a small amount of data can be
sent over the connection, or (2) the connection can be cleanly shut
down. ConnectionHealthChecker may be provided multiple remote server
addresses, and it attempts them in a round-robin fashion if one attempt
fails.

BUG=chromium:225913
TEST=Build and run unit tests.

Change-Id: I38d9c107f4dd82ea780ee628d0b8bf4ec4929cf0
Reviewed-on: https://gerrit.chromium.org/gerrit/47248
Reviewed-by: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>