shill: Renames Config80211 to NetlinkManager.
The domain of Config80211 has extended to different types of netlink
message communication.  That warrants a name change.  This CL changes
the name of the class plus a bunch of objects of that class.  This
required changing the names of some files.

In addition, a few (half a dozen, maybe) readability issues were
addressed.  These include inaccurate header guard comments and bad
indentation.

BUG=None
TEST=unittest

Change-Id: Id0006d5d37a98ed6c138f9982a2aa3624b308757
Reviewed-on: https://gerrit.chromium.org/gerrit/47860
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Wade Guthrie <wdg@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>