shill: EapCredentials: Promote to a class
Promote the EapCredentials structure to a sovereign class with
a small set of getters and setters where necessary.  Otherwise,
move most of the member manipulation (creating arguments for
wpa_supplicant and wimax_manager, Load() and Save()) interior
to the class.  This involves a straightforwarnd but lengthy
refactor of just about everything that uses EapCredentials.
The EapCredentials object is ref-counted to allow it to be
shared between some services.  For this reason, move the
"remote_certification" property out of EapCredentials into
Service since this property is strictly per-service.

BUG=chromium:227434
TEST=Unit tests, wired 802.1x and wifi EAP-TLS autotests

Change-Id: I167e71e1d9c934bc328a0e1d73f638e1ff50d238
Reviewed-on: https://gerrit.chromium.org/gerrit/47973
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>