shill: WiFiService: Remove errant warnings/errors on load
Fix two misleading warnings on loading a WiFiService.  First,
the warning "More than one configuration for service..."
appears on all successful loads due to an improper inequality
check.  Second, if an empty passphrase is loaded (or none at
all), we should not emit an error if passphrases are not
supported for this service type.

BUG=chromium:230318
TEST=Unit tests

Change-Id: I4724f3399e7eee096922e94b669d2798391c0912
Reviewed-on: https://gerrit.chromium.org/gerrit/47855
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>