shill: Adds capability to supply custom parser for netlink attributes.
Some netlink attributes (notably BSS information elements) are provided
as a blob of raw data.  Since the parsing of these blobs is specific to
the attribute in question, the best way to deal with this is to tie the
parser directly to the attribute.  This CL provides a mechanism to do
just that.

BUG=chromium:221118
TEST=unittest (plus, there're some CLs, coming up, that use this feature).

Change-Id: I480cc586e48465f5c8ded816b84e5723097df20a
Reviewed-on: https://gerrit.chromium.org/gerrit/48012
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>
Commit-Queue: Wade Guthrie <wdg@chromium.org>