shill: simplify log messages during make
At present, we get tons of log spew, which makes it hard to find
error messages in the terminal buffer. Fix this by a) removing
-print-cmdline, and b) suppressing the commands run by Makefile.

To keep the build process from being completely silent, we
manually echo an abbreviated name for each command run, and the
target it is building. We don't, however, output the arguments
to the command.

While there: replace a use of cat with cp (makes the recipes
more uniform).

BUG=chromium:233843
TEST=make clean; cros_workon_make --test --board=link --reconf shill

Change-Id: Ie5fbe491bcdf8005bc3ae31d2a7aed7a58b33542
Reviewed-on: https://gerrit.chromium.org/gerrit/48735
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>