shill: Only call shim result callback from clean shim state
Fix a bug where VerifyAndEncryptData fails when the Verify callback
cannot start an Encrypt operation because the shim is supposedly still
working.  The shim has actually returned a complete result in this case,
but we're calling into the crypto_util_proxy before giving it a chance
to clean up the shim state.

While we're here, check the correct error in the Manager when we fail to
link Verify and Encrypt shim operations.

TEST=Unit tests.
BUG=chromium:236083

Change-Id: I00366d55afa1c3dd228d40662e2baacac7d79eb0
Reviewed-on: https://gerrit.chromium.org/gerrit/49490
Commit-Queue: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>