shill: fix default settings for new profiles
While fixing chromium:221021, I broke portal detection (and a couple other
features). This happened because, in the absence of an existing default
profile, shill would create a default profile, and save it immediately. This
circumvented the logic in DefaultProfile::LoadManagerProperties, which
initializes the default profile with default values for some properties.

Fix this by restoring the call to Manager::LoadProperties, in
Manager::InitializeProfiles.

BUG=chromium:237927
TEST=new unit tests, manual

Manual test
-----------
- # stop shill
- # rm /var/cache/shill/default.profile
- # start shill
- # grep 'CheckPortalList=ethernet,wifi,cellular' /var/cache/shill/default.profle
  -> should show a line exactly matching the pattern

Change-Id: I4696ce445b45cad05a167d91c41e8f5eb486fe0c
Reviewed-on: https://gerrit.chromium.org/gerrit/50081
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>