shill: Times-out old netlink response handlers.
Adds a default timeout to all message response handlers.  Adds a 'type'
to auxilliary message handlers.  When/if a netlink message goes
unanswered (we don't receive a unicast response with the same serial
number) after the timeout (and there's no intervening sent message -- we
only timeout old handlers on message send), the axilliary message
handler is called with a 'type' of 'timeout' and the handler is deleted.

BUG=chromium:249028
TEST=unittest

Change-Id: I85e84a27acfaa6d9b6af6f25288056bda40a649b
Reviewed-on: https://gerrit.chromium.org/gerrit/59306
Commit-Queue: Wade Guthrie <wdg@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>