shill: fix profile-api documentation, remove stale comments in dbus_bindings
Profile.DeleteEntry was documented as expecting a D-Bus path,
when, in fact, it takes a string.

While there, remove a stale comment about auto-generating the XML
files, and installing them to a system directory. The first can't
be done, because this XML is how we generate RPC stubs in the first
place; the second is already done in the Makefile.

BUG=None
TEST=emerge-link shill

Change-Id: I3dbd41ece19cbda147bf894f66b5b29a47b7651a
Reviewed-on: https://gerrit.chromium.org/gerrit/59458
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>