shill: EAPCredentials: Use PEM arrays for CACertPEM
Switch the "CACertPEM" option to use a string array of PEM.
wpa_supplicant is equally capable of accepting PEM and DER
certificates for its "ca_cert" option.  Since it's not easy
to export multiple certificates in a single DER file, use
a PEM file instead of DER for the output file in the CaCertPEM
case.

BUG=chromium:249363
TEST=Unit tests.

Change-Id: I3dd8389a3b099a368924459521f4978effe12b7a
Reviewed-on: https://gerrit.chromium.org/gerrit/60522
Commit-Queue: Paul Stewart <pstew@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>