shill: Connection: Don't create a route to invalid gateway
Automated tests like network_VPNConnect will create VPN links
on interfaces that may not have a gateway route.  This causes
a crash in shill since it does not ever expect the underlying
connection to have no gateway.  Although this is unlikely in
the field, there is no cost to fixing this issue in shill.

BUG=chromium:259030
TEST=Unit tests

Change-Id: I25372df5066ad4c044fdf542f83934ab1418f517
Reviewed-on: https://gerrit.chromium.org/gerrit/61450
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>