shill: Makes progressive scan field trial more visible in logs.
The progressive scan field trial used to be shown in the logs only on
instantiation of the WiFi device so feedback would probably not provide
this information.  This CL makes the progressive scan group visible at
the start of every scan greatly increasing the likelihood of having this
data available in feedback logs.

BUG=None
TEST=unittest

Change-Id: I5c6d7b563002339740ae54e919843f218cd1ba8e
Reviewed-on: https://gerrit.chromium.org/gerrit/63360
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>
Commit-Queue: Wade Guthrie <wdg@chromium.org>