shill: dongles: fix handling of ppp failures
There was an asymmetry in how we were handling PPP failures.

If pppd exited without first reporting a disconnect event,
as would happen in the case of "kill -9", we would Disconnect()
the underlying Cellular connection. This would cause shill
to retry the connection, as desired.

However, if pppd reported a disconnect event, then
OnPPPDisconnected would not Disconnect(). This left things in a
state where Cellular was connected, but ppp was not running.

You might ask why kill -TERM didn't cause OnPPPDied to Disconnect()
the Cellular connection. The answer is that OnPPPDisconnected
discards the ExternalTask which is responsible for invoking
OnPPPDied.

An alternative fix would be to make OnPPPDisconnected informative,
rather than conclusive. That is, we could leave ppp_task_ alone at
that point, under the assumption that pppd is going to retry the
connection, and (eventually) time out and exit.

But in the case where PPP failed due to authentication issues, we
should terminate pppd. And as long as we're doing that for the
authentication failure case, it seems simpler to handle other
failures in the same way.

Note that the above argument means that we shouldn't reconnect
when we have bad credentials. I'll fix that in the next CL.

BUG=chromium:246451
TEST=unit tests, manual

Manual test
-----------
1. plug in PPP dongle
2. wait for modem to connect
3. killall -9 pppd
4. wait (< 1 min) for modem to re-connect
5. killall pppd
6. wait (< 1 min) for modem to re-connect

Change-Id: I6aa084414a5bbba8bb57c9bbf885adc13596fa54
Reviewed-on: https://chromium-review.googlesource.com/65918
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>