shill: dongles: don't retry connect after PPP auth failure
An earlier CL for chromium:246451 fixes a problem where we might not
reconnect after a PPP failure. However, that changes the behavior
for the authentication failure case. Whereas we would previously
not reconnect in that case (just as in any other case), we now try
reconnecting.

Automatically reconnecting after an authentication failure doesn't
make sense, since we expect the same result on the next attempt.
Fix this by having CellularServices return false for IsAutoConnectable,
when in this state.

The user should change the PPP credentials, and then initiate a connection
manually.

While there: remove an unused declaration in service.h.

BUG=chromium:246451
TEST=unit test, manual (pending)

Manual test
-----------
- stop shill
- rm /var/cache/shill/default.profile
- rm /var/run/shill/user_profiles/chronos/shill.profile
- start shill
- plug in dongle that requires a ppp username and password
- watch connection fail, with notification about PPP credential failure
- wait (e.g. 10 seconds), verify that shill does not retry connecting
- configure correct username and password
- select cellular network from ash tray
- verify that we connect to the network
- suspend + resume
- verify that we reconnect to the network

Change-Id: I92b3b9048af33731c289f1f1d29e125a557a496a
Reviewed-on: https://chromium-review.googlesource.com/65919
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>