shill: remove Network.Shill.CellularDropsPerHour metric
This metric can't be too important, because nobody noticed it was
missing from histograms.xml (and hence, not viewable in the metrics
dashboard).

Also, a per-hour metric can be misleading, because it doesn't account
for what fraction of the hour the device was actually trying to use
Cellular.

This commit sort-of reverts 6c1e3bbc, but it leaves some of the
hooks of that commit in place. In particular, metrics still takes
an EventDispatcher pointer in its ctor, so that it can report
periodic statistics. While there aren't presently any other periodic
statistics, it's easier to leave these hooks in place than to rip
them out.

BUG=chromium:274640
TEST=unit tests

Change-Id: I7ad88e2e2dd54935f4d8f743203802612eed52df
Reviewed-on: https://gerrit.chromium.org/gerrit/66123
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>