shill: wimax: Do not clear EAP credentials upon connection failures.
This CL changes the code not to clear the EAP credentials upon
connection failures, which prevents the UI from prompting the user for
the EAP passphrase unnecessarily when the connection fails due to low
signal coverage. However, that also prevents the UI from prompting the
user if an incorrect EAP passphrase is given, which needs to be
addressed in a follow-up CL. This CL also adds more WiMAX-related debug
logging.

BUG=chrome-os-partner:22162
TEST=Build and run unit tests.

Change-Id: Iee2d9f157267b7c3e9fa68143fa446f827a5dbba
Reviewed-on: https://chromium-review.googlesource.com/66889
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>