shill: Clean up Service::Remove()
This method used to leave the service in an inconsistent state
(unloading the service while still linking it with a profile that
contained its configuration).  This CL adds a real implementation
which disassociates the service from its current profile and
unloads the service if no other profile supports it.  It also
supports removing the service from the Manager's list if the
service should no longer exist.

BUG=chromium:281655
TEST=Unit tests + manual:
Call RemoveService on a visible, remembered service and note that
the service is unloaded and no longer associated with a profile.
Call RemoveService on a remembered but non-visible service and
note that the service is unloaded, the configuration no longer
exists and the service is removed from the Manager ServiceCompleteList.

Change-Id: I985c4498b5acf55b8f3d78ee570d30597a2c5384
Reviewed-on: https://chromium-review.googlesource.com/167530
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>