shill: cellular: CellularCapabilityGSM: No IMSI doesn't mean no SIM.
CellularCapabilityGSM uses the availability of IMSI to determine if a
SIM card is present. This logic fails if the SIM is locked, as the SIM
will be present but the IMSI won't be accessible due to the lock. This
CL fixes the code to take SIM lock into account.

BUG=chromium:290549
TEST=Build and run unit tests.

Change-Id: I6cffcf97e93e1361b28b094974825eb9910ee770
Reviewed-on: https://chromium-review.googlesource.com/169240
Reviewed-by: Arman Uguray <armansito@chromium.org>
Commit-Queue: Arman Uguray <armansito@chromium.org>
Tested-by: Arman Uguray <armansito@chromium.org>