shill: Combines duplicate scan field-trial configurations.
The 4/4/50% configuration is a duplicate of the 4/4/33% configuration
since the percent parameter doesn't matter when min==max. Similarly, the
8/8/100% configuration is a duplicate of the 8/8/50% configuration.

Fixing the duplicates to avoid confusing our future selves.

Removing them entirely might lead to confused data in the UMA dashboard
so we're just folding the 50% config into the 33% config, and the 100%
config into the 50% config.

BUG=chromium:299999
TEST=unittest

Change-Id: Ie8f29a31f1d93003ccf8d6fc686067de4fd249b2
Reviewed-on: https://chromium-review.googlesource.com/171676
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Commit-Queue: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>