shill: cellular: Use OTA info for APN list if SIM is not readable.
This CL changes the behavior so that if SIM info is unknown, the OTA
data will be used to construct the APN list. This is done by assigning
the OTA code as the home provider code and setting the home
provider to be the same as the serving operator. This is to remedy an
issue that prevents modems from connecting if the SIM information is
unavailable.

BUG=chromium:306310
TEST=On a spring with LTE do:
     1. Send AT+CFUN=4 to modem.
     2. Restart ModemManager.
     3. Use mmcli -i <sim #> to verify that all info from the SIM is unknown.
     4. Use mmcli -m <modem #> to verify that the 3GPP properties have
     updated to display a valid opertor code and operator name.
     5. Use connectivity show devices to verify that the cellular device
     has a populated Cellular.APNList property.

Change-Id: I6fe713b986e9b2b4767410d545941aa7e1ab9915
Reviewed-on: https://chromium-review.googlesource.com/172941
Reviewed-by: Arman Uguray <armansito@chromium.org>
Tested-by: Arman Uguray <armansito@chromium.org>
Commit-Queue: Arman Uguray <armansito@chromium.org>