shill: Service: Remove "Favorite" property
We bid farewell to the much misunderstood and maligned "Favorite"
property.  Chrome doesn't use this property and although it did
not fall into disuse in shill, it was misused for the purpose of
tracking whether the "auto_connect" property should be modified
on a successful connection.  This CL renames the property and its
accessors more intuitively.  We will no longer export and emit
property change events for this property.

While here, remove "Favorite" from the properties used to compare
services for sorting.  Replace it with a test for HasEverConnected.

BUG=chromium:308754
TEST=Unit tests

Change-Id: I577849cbcff2758b02e65516eafc3f9e5c20e888
Reviewed-on: https://chromium-review.googlesource.com/173631
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: David James <davidjames@chromium.org>