shill: DHCPConfig: Get new lease with RenewIP
If a DHCP process is not running when RenewIP is called, acquire
a new lease instead of returning.  This allows callers external
to shill to avoid having to keep track of whether the IPConfig
object has a lease or not.

BUG=chromium:318290
TEST=Unit test + manual: Reproduce the issue in crbug.com/318290.
In specific, change the IP address to a static IP first and close
the configuration panel, then separately switch to custom name
servers.

Change-Id: Ifebf7507f65dc981e5ca77d000ba298e0c6f22ee
Reviewed-on: https://chromium-review.googlesource.com/177350
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>