shill: Leave DBus watcher for wpa_supplicant in place on Stop()
By resetting our reference on Stop(), we are triggering a
SupplicantAppeared event when we re-enable the WiFi device, which
restarts the device.  Restarting the device however calls Stop() again,
and the cycle repeats.  The WiFi event handlers for supplicant vanishing
and reappearing expect notifications even if the device is disabled.  We
are safe to leave the watcher in place for the lifetime of the WiFi
device.

BUG=chromium:334848
TEST=With this patch, shill no longer enters a restart loop when
WiFi is disabled and then later re-enabled.

Change-Id: Iee87000e78504ccc233f066c1ac7a00b532fe82e
Reviewed-on: https://chromium-review.googlesource.com/182735
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Commit-Queue: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>