shill: cellular: Remove termination action in Cellular destructor.
Cellular registers a termination action with the Manager when the device is
enabled. This action is currently removed in the Cellular::StopModemCallback.
This callback is not guaranteed to run when the underlying modemmanager modem
object disappears. This CL also removes the action in Cellular destructor to
avoid leaking the termination action after the Cellular object has been
destroyed.

BUG=chromium:304557
TEST=(1) On a device without a modem, run the following steps:
         - stop modemmanager; stop cromo
         - restart shill
         - /usr/local/autotest/cros/cellular/pseudomodem/pseudomodem.py
               --family '3GPP'
         - kill pseudomodem. (Ctrl-C on the same terminal)
         - run power_Resume. It should pass.
     (2) Sanity check by running network_3GModemControl.

Change-Id: Id4cb70414da805ac2f7d60b3670a05c7e059eb32
Reviewed-on: https://chromium-review.googlesource.com/184270
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Thieu Le <thieule@chromium.org>