shill: WiFi: Support driver-based roaming
When we switch between BSSes of a connected service, we should
renew our DHCP lease in the unlikely event that this network
is on a different subnet.  Currently this works when roaming is
performed by wpa_supplicant (in network_WiFi_Roam) since it
is currently rare that we succeed in roaming before shill
intervenes and restarts the connection from scratch.

When the device / firmware performs the roam, neither
wpa_supplicant nor shill may be notified during the disconnect
event, so the first event visible to shill may only be the change
in CurrentBSS.  As a result, the service never disconnects.  In
this situation we should also renew our DHCP lease.  This
CL detects such roaming behavior and triggers a DHCP renewal.

BUG=chromium:342097
TEST=Autotest wifi_roamNone

Change-Id: If581fa755ea3d241e221b4906f873694cd51b40e
Reviewed-on: https://chromium-review.googlesource.com/185799
Reviewed-by: <dnkim1205@gmail.com>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>