shill: Service: Track previous error details
Add new PreviousError and PreviousErrorSerialNumber properties
to the Service so that the UI can query the last value of the
Error property even after this failure is cleared (and the
Service may no longer be in an errored state).

BUG=chromium:302020
TEST=Addition to unit tests + manual: list-service and observe
that the new properties are displayed.

Change-Id: Ia051908114c8821f173f2d2962189b8785c1da69
Reviewed-on: https://chromium-review.googlesource.com/187069
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>