shill: re-order notifications in Manager::SortServicesTask
Handle processing of a new Service before sending out notifications
about changes in Services and Technologies.

The primary difference this causes is that resolv.conf will be
written to disk (via Connection::SetIsDefault) before we notfiy
external parties of the change in service ordering, which technologies
are connected/default.

Note that while this _might_ help with chromium:330873, the better
fix for that is in chromium:159540. I think this change is worth making
even in the presence of the fix for chromium:159540, because the
code reads cleaner this way. (We now update all of our local state
before generating external notifications.)

BUG=chromium:345155
TEST=unit tests

Change-Id: Iba3861e6f3334437e03e755971cfacd59a7c17fb
Reviewed-on: https://chromium-review.googlesource.com/187885
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>