shill: Crash immediately when failed to create necessary DBus adaptors.
A few classes in shill creates a DBus adaptor at construction and freely
accesses the adaptor object under the assumption that the adaptor can
always be created successfully. In reality, we have observed crashes due
to accessing a NULL adaptor. However, as the crashes happened at various
call sites when a NULL adaptor was accessed, the crash rate of each
intance is low, which makes it easy to miss the bigger problem where
shill created objects with duplicated identifiers.

This CL makes sure that the code crashes immediately when it fails to
create a necessary DBus adaptor so that we can obtain the same crash
signatures for scenarios where shill fails to create a DBus adaptor.

BUG=chromium:350188
TEST=Build and run unit tests.

Change-Id: I9436152d51de2968c6ea92fa749dd200d0633852
Reviewed-on: https://chromium-review.googlesource.com/189116
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>