Improve disconnect reporting from shill
When service is disconnected or unable to connect, set the failure state
to an appropriate value if failure state is not set yet and disconnect is
not initiated by user, and then set the service state to idle.

When service is disconnected due to intentional network switch, do not
set the failure state, only set the service state to idle.

BUG=chromium:352383
TEST=manual
out-of-range test
1. Connect a target device to an AP.
2. Manually turn off the AP.
3. Verify in Network Event Log, the service state is first transition to
   failure (out-of-range error), and then idle.

bad passphrase test
1. Trying to connect to an AP with wrong passphrase.
2. Verify in Network Event Log, the service state is first transition to
   failure (bad-passphrase error), and then idle.

network switching test
1. Connect a target device to an AP "AP1"
2. Manually connect the device to another AP "AP2"
3. Verify in Network Event log, the service state for "AP1" is transition
   to idle.

Change-Id: I4141dbe5de5225ccd6e3e1a64d8f61b31009b9d9
Reviewed-on: https://chromium-review.googlesource.com/190119
Tested-by: Peter Qiu <zqiu@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Peter Qiu <zqiu@chromium.org>