shill: downgrade log message from WiFi::BSSRemovedTask
Because we don't create Services or Endpoints for SSIDs that are empty
or that start with the NULL byte, a BSSRemoved message for such a BSS
causes us to log a warning.

However, BSSes with empty SSIDs are common, as that is how hidden
SSIDs are implemented. As such, this condition doesn't really merit
a warning.

Accordingly, downgrade the message to VERBOSE1.

BUG=chromium:352822
TEST=unit tests

Change-Id: I1a63e46faf34970d9bdc2c90a9494f8fec60dbb1
Reviewed-on: https://chromium-review.googlesource.com/190443
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>