shill: enable wake-on-lan for ethernet
For Ethernet devices, enable Wake-on-Lan via Magic Packet,
and disable Wake-on-Lan via other methods.

For the Linksys USB300M, this reduces the set of WOL
triggers. (The USB300M defaults to PHY and Magic Packet).

For r8169, this is essentially a no-op. The default for r8169
(at least on Panther) is to enable WOL via Magic Packet
(and no other methods).

While there: change EthernetTest to use NiceMock<MockMetrics>
instead of plain MockMetrics. (This reduces log-spew, and
we weren't setting any expectations on the mock anyway.)

BUG=chrome-os-partner:25273
TEST=manual (see below)

Manual test
-----------
- plug linksys usb300m ethernet dongle in peppy
- # ethtool eth0 | egrep '^\s+<Wake-on'
  Wake-on: g
- The default for this device is 'Wake-on: pg'.
  So if you use 'pg' instead of 'g', then verification has
  failed.

Change-Id: I310167fa2d89996500c12100f68c28420b1bb963
Reviewed-on: https://chromium-review.googlesource.com/192665
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>