shill: cellular: Fail some MobileOperatorInfo unittests early.
In two places, the unit-tests was trying to compare details of an object that
had been shown to not exist (in which case the test fails anyway). This caused a
segfault in the unit-test. By promoting the Expect* to Assert* for these
instances, we quit the test early.

BUG=None.
TEST=Run shill unittests.

Change-Id: Iad84d96fd111bf21e96f83db175dabbe78d52c2f
Reviewed-on: https://chromium-review.googlesource.com/195074
Reviewed-by: Ben Chan <benchan@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>