shill: cellular: Don't explicitly write to service profile on property updates.
The |Service| class has hooks into the property store that allow it to update
the non-ephemeral profile associated with a service whenever a property update
occurs.
This CL removes some explicit writes to the profile in CellularService. This
writes had a side-effect of assigning a non-ephemeral profile to the service.
Instead, we should wait till the |Service| class assigns a non-ephemeral profile
to the service.

BUG=chromium:364633
TEST=Manually tested that
     (1) On a successful connection, the LastGoodApn is written to the service
         profile.
     (2) When the user enters a custom apn from the UI, it gets written to the
         service profile.

Change-Id: Iadc35e73b3f28c984d2b02f4796b5e02c246facb
Reviewed-on: https://chromium-review.googlesource.com/196400
Reviewed-by: Thieu Le <thieule@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>