shill: DeviceInfo: Report IPv6 address changes to Device
In the current scheme of operation, shill is responsible for
launching negotiation and configuring IPv4 addresses into the
kernel.  However the kernel is responsible for negotiating and
configuring IPv6 addressess via neighbor discovery.  This CL
provides a means for DeviceInfo to report IPv6 address changes
back to the Device.  The Device can then choose to report this
state over the RPC interface.

BUG=chromium:365883
TEST=Unit tests

Change-Id: I9b25891907e2bb160b82ea2449fd225d4967c6bc
Reviewed-on: https://chromium-review.googlesource.com/196960
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>