shill: cellular: Migrate activation paths to use MobileOperatorInfo.
The code paths that determine whether cellular activation is required also
verify that some online portal information is available. This CL migrates this
check to be based off of the new MobileOperatorInfo objects. Before this CL, a
new |CellularOperatorInfo| object was used to make this decision.

BUG=chromium:371630
TEST=- Use pseudomodem to manually create a modem that exposes itself with
         operator id: '310995'
         operator name: ''
         subscription state: 'unknown'
         own: '0000000000'
       and verify that
         $ connectivity show devices
       shows the cellular service is 'not-activated'.
     - Run network_LTEActivate.

Change-Id: I22cecffca1169e8abf25148b367d01520aa171bc
Reviewed-on: https://chromium-review.googlesource.com/200687
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>
Reviewed-by: Thieu Le <thieule@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>