shill: WiFi: Disable gateway ARP for static IP
Even when a static IP address is configured for a WiFi service,
we perform a DHCP lease negotiation in the interest of gaining
other parameters associated with the connection (DNS servers,
proxy configuration, etc).  Shortly thereafter, we release such
a lease.  Since we never intend on actually using the IP address
provided by the DHCP server, the validity of the IP address it
provides is not of any interest.  As such, it's reasonable to
disable the "gateway ARP" address validation.

BUG=chromium:377990
TEST=New unit tests.  Note that an DHCP autotest using
pseudoethernet won't help validate this change since it is not
WiFi, and as such gateway ARP is disabled across-the-board.

Change-Id: I7f3e096370189698f900f85d78592fd42017ffbd
Reviewed-on: https://chromium-review.googlesource.com/202176
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>