shill: WiFiService / WiMaxService: Emit Visibility updates
Send property change singals for the "Visible" property when
some attribute of the service changes in a way that would
affect it.  For WiMax services, this happens during start/stop
and for WiFi services this happens when the endpoint list or
service state changes.

Also add documentation for the "Visible" property.

BUG=chromium:379231
TEST=Unit tests

Change-Id: I104226fe096996a58c251984af498ebe1910a120
Reviewed-on: https://chromium-review.googlesource.com/202136
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>