Pending service connection attempt.
Disconnect the pending service when wpa_supplicant is connected to
a service that's neither pending or current. We get into this state
when we have two connection attempts to two different services in a
quick succession, where the second attempt failed because of the
cfg80211 driver is still busy with the first connection attempt. Thus,
the connection attempt to the second service is ignored by wpa_supplicant.

By disconnecting the pending service in this situation, it will send a
"Disconnect" command to wpa_supplicant, which will disconnect the unintended
service. It will also mark the pending service as connectable (not connecting),
and kick off the connection attempt by manager for the pending service sooner
rather than waiting for the pending timeout.

BUG=chrome-os-partner:29436
TEST=unit tests, run network_WiFi_ProfileBasic

Change-Id: I7faeeb42a324548e5b2df892781dd6e668cdf0fe
Reviewed-on: https://chromium-review.googlesource.com/203147
Tested-by: Peter Qiu <zqiu@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Peter Qiu <zqiu@chromium.org>