shill: cellular: Special case operator matching for CUBIC.
We do not have the MCCMNC values associated with this operator. Since it is an
MVNO, it shares MCCMNC values with other operators, which leads us to make an
incorrect match with the other operator instead of CUBIC.

This CL adds a stop-gap solution where we special case CUBIC to match before we
look at the MCCMNC values at all.

BUG=chrome-os-partner:30200
TEST=(1) Using pseudomodem, tested that CUBIC is correctly identified even when
         sharing MCCMNC value with another MNO.
     (2) Tested that we can connect with the CUBIC network.

Change-Id: I4bc77314c49d0f2be3df9e62396936ee5bf74f9b
Reviewed-on: https://chromium-review.googlesource.com/206754
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>