shill: cellular: Provide IMSI / ICCID values from the SIM to serving operator.
Although we obtain IMSI and ICCID values only from the SIM, we would like to use
them in the identification of MVNO that is the current serving operator.
This CL wires up the Capability* classes to provide that information to the
serving operator.

BUG=chrome-os-partner:23086
TEST=TODO(pprabhu) crbug.com/371027 Write an autotest test that mimicks the
network condition to test the system level behaviour wrt MVNO identification.

Change-Id: I78027a4adbb4e8499bfbd40f1883f5f91e21c2c7
Reviewed-on: https://chromium-review.googlesource.com/210049
Reviewed-by: Ben Chan <benchan@chromium.org>
Commit-Queue: Prathmesh Prabhu <pprabhu@chromium.org>
Tested-by: Prathmesh Prabhu <pprabhu@chromium.org>