shill: Fix Network.Shill.UserInitiatedEvents histogram
Reserve an unused bucket for UserInitiatedEvents histogram to make it
a non-single bucket histogram. Chromium does not allow single
bucket histogram with bucket value starting at 0, since it requires
minimum of 3 buckets per histogram including the underflow and the
overflow buckets.

Bucket with value of 0 is the underflow bucket, and bucket
with value of |max| is the overflow bucket. There would only be two
buckets when single bucket histogram starts at 0 (bucket 0 and 1).

So this fix avoids any change to the existing bucket value for
UserInitiatedEvents histogram in histogram.xml. This fix also allow the
existing bucket value to stay the same when adding more buckets to this
histogram in the near future.

BUG=chromium:369545
TEST=unit tests, manual
1. On a chrome device, open the network menu in chrome to initiate
   a wifi scan.
2. Browse to "chrome://histograms" to verify there is a histogram for
   "Network.Shill.UserInitiatedEvents", and a hit for value 0 (WifiScan).

Change-Id: Ic0d0c6d91dd14a5dcf6c99ee291cd630781c7018
Reviewed-on: https://chromium-review.googlesource.com/211261
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Peter Qiu <zqiu@chromium.org>
Tested-by: Peter Qiu <zqiu@chromium.org>