shill: Service: Load has_ever_connected_ properly on profile load
The stored value for has_ever_connected was loaded before loading
credentials, resulting in the value being false.  This CL changes this
logic to load the value after loading the credentials.  To avoid the
side-effects of the OnEapCredentialsChanged() call resulting from the
profile credential load, the has_ever_connected_ variable is set to
false.  This is documented with a comment in the code.  The
service Load test was updated to detect a regression and an additional
test was added to verify that a change to the value would be stored in
the profile when the SetHasEverConnected(bool) function is called and
the profile is saved.

BUG=chromium:401193
TEST=Ran shill unit tests

Change-Id: Ia8852f3aa6a8fc9fa56e1674be966e0da0c95e21
Reviewed-on: https://chromium-review.googlesource.com/211521
Tested-by: Rebecca Silberstein <silberst@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Rebecca Silberstein <silberst@chromium.org>