shill: Create ConnectivityTrial class to hold logic for connection tests
Create a new class, ConnectivityTrial, that contains the logic and handling
for each attempt in portal detection.  This separates the higher level logic
of retries and maximum number of attempts from the individual connection tests.

This architectural change enables code-reuse for an upcoming feature for on
demand connection testing.

BUG=chromium:402888
TEST=ran shill unit tests and tested on peppy

Change-Id: I90d2bca4cd0778c3aeac82677b675247769a5597
Reviewed-on: https://chromium-review.googlesource.com/216505
Commit-Queue: Rebecca Silberstein <silberst@chromium.org>
Tested-by: Rebecca Silberstein <silberst@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>