shill: Add Service property "SavedIPConfig" and "StaticIPConfig"
Both "SavedIPConfig" and "StaticIPConfig" contains a dictionary of
IP parameters. NameServers field within the dictionary is an array
of string instead of comma separated string.

This will make the saved IP and static IP properties more consistent
with the IPConfig object properties. The legacy properties
("SavedIP.*" and "StaticIP.*") can be removed once the Chrome switch
over to use these new dictionary.

Note, the names and formats of the static IP configuration saved
in the storage are not changing. This is to avoid any backward
compatibility issues.

BUG=chromium:411289
TEST=unit tests, manual test
Manual Test:
1. Connect a chrome device to "GoogleGuest", run "./list-services"
   from command prompt, verify "SavedIPConfig" field for service
   "GoogleGuest" contain the dictionary of IP Parameters.
2. Manually configure "GoogleGuest" with IP address and Name Servers,
   run "./list-services" from command prompt, and verify "StaticIPConfig"
   field for service "GoogleGuest" contain the dictionary of IP parameters
   with manual configured values.
CQ-DEPEND=CL:217115

Change-Id: I6741c80d3889f15319736d34d3ca466e36874983
Reviewed-on: https://chromium-review.googlesource.com/217135
Reviewed-by: Peter Qiu <zqiu@chromium.org>
Commit-Queue: Peter Qiu <zqiu@chromium.org>
Tested-by: Peter Qiu <zqiu@chromium.org>