shill: detect password change on user profile load
When loading a user profile, the stored password should be checked
against the currently active password.  A credential change should
be detected and handled if the passwords are different.  Additionally
removed the Service::SaveToProfile call in Service::SetHasEverConnected
to avoid overwriting a previously loaded profile.  Removed ordering
dependence on loading parts of a profile by adding a boolean to
Service::OnEapCredentialsChanged to differentiate between loading
credentials from a profile and changing credentials. In WiFi services
that do not use EAP for authentication, ignore changes to EAP credentials.
This allows only passphrase changes to trigger a disconnect when the
service is loaded from a new profile.

BUG=chromium:414889, chromium:414562
TEST=unit tests with and manual verification on peppy:
Connect to a new PSK network while logged in.  Log out (system should
disconnect).  Connect to the same network again (while still logged out).
Log in, and ensure that the system does not disconnect and re-connect to
the network in the process.

Change-Id: I6eef098728951cfd0118e94278d76caf595d5b81
Reviewed-on: https://chromium-review.googlesource.com/218618
Reviewed-by: Rebecca Silberstein <silberst@chromium.org>
Tested-by: Rebecca Silberstein <silberst@chromium.org>
Commit-Queue: Rebecca Silberstein <silberst@chromium.org>