shill: Disable HT40 when WiFi link becomes unreliable
OBSS scan will not be performed when HT40 is disabled, which will
reduce the load on the client. And hopefully, would help improve the
link quality.

Once HT40 is disabled for a network, it will only be re-enabled when
system resumed from suspend or when new connection attempt to that
network is initiated.

BUG=chromium:419046
TEST=unit tests, manual test
Manual Test:
  Hardcode the code to inject link monitor failures and verify
  through debug messages.
CQ-DEPEND=CL:220651

Change-Id: I87400a0628228679a935b4b0455738d316a16627
Reviewed-on: https://chromium-review.googlesource.com/221296
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Peter Qiu <zqiu@chromium.org>
Commit-Queue: Peter Qiu <zqiu@chromium.org>