shill: Service: Rank HasEverConnected with Managed Credentials
Add a service property that specifies whether the credentials
for this service are managed outside of shill, and are
validated in some manner other than the user.  As such, these
credentials need not be validated via an initial connection
attempt in order for the service to be considered of equal
priority to other services that have been connected before.

BUG=chromium:424036
TEST=Unit tests

Change-Id: I20f709b525a32400ae9b90797f8107c2f6446f2e
Reviewed-on: https://chromium-review.googlesource.com/223653
Reviewed-by: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>