shill: Manager: Add ProhibitedTechnologies property
Add a Manager property for the list of technologies that should
be prohibited from enablement through D-Bus APIs.  New devices
that appear will similarly start in the disabled state and cannot
be enabled.  Removing a technology from the prohibited list does
not re-enable any devices that were previously disabled.

CQ-DEPEND=CL:231738
BUG=chromium:436336
TEST=Unit tests
Also manual:
  dbus-send --system --print-reply --fixed \
       --dest=org.chromium.flimflam / \
       org.chromium.flimflam.Manager.SetProperty \
       string:ProhibitedTechnologies \
       variant:string:wifi

At this point, WiFi is disconnected and disabled, and cannot
be re-enabled via the UI.  Changing "wifi" to "wimax" above
causes it to be possible to enable WiFi in the UI again.

Change-Id: I2e95cb5e95913739ccd54457bec778951af4dae6
Reviewed-on: https://chromium-review.googlesource.com/231739
Reviewed-by: Ben Chan <benchan@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>