shill: check connectivity before scanning on resume
Change WiFi::OnAfterResume logic such that we only initiate a
scan on resume if we are not connected.

In the corner case where we appear to be connected on resume but
are actually on our way to disconnecting, this change will slow
the reconnect time since a scan will only be triggered some time
after WiFi::OnAfterResume is executed. However, this prevents
a background scan from being launched if we are indeed connected
on resume, which decreases WiFi throughput and increases latency.

BUG=chromium:438831
TEST=Compile shill and run unit tests using:

P2_TEST_FILTER="shill::*" FEATURES="test" USE="clang asan" \
emerge-samus shill

Change-Id: Ia7aed7ef7c1ff6f993d2200d8b97b2787ed718d2
Reviewed-on: https://chromium-review.googlesource.com/233033
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Samuel Tan <samueltan@chromium.org>
Tested-by: Samuel Tan <samueltan@chromium.org>