shill: Ethernet: Retain EthernetService for life
Make the EthernetService held by the Ethernet device remain valid
for the lifetime of the Device.  This allows Ethernet services to
be configured whether or not a link is available.  To modify this
lifetime, it is necessary to break the reference loop between the
Ethernet device and its Service.

Instead of calling RegisterService and DeregisterService with the
EthernetService, toggle the visibility and autoconnect properties
to allow the service to disconnect and regain L2 connectivity as
link appears and disappears.

BUG=chromium:439176
TEST=The following:
 - Unit tests with ASAN
 - The following autotests:
   + network_Dhcp*
   + network_VPNConnect
 - Manual:
   + Start shill with Ethernet connected, disconnect, reconnect
   + Start shill with Ethernet USB disconnected, connect USB (with Ethernet attached)
   + Start shill with Ethernet USB disconnected, connect USB with Ethernet detached,
     then connect Ethernet
   + Monitor "list-services" to make sure the Ethernet service is only visible
     (and connects) when both USB and Ethernet are connected.

Change-Id: I7f0ef8397606b69990b8dc03be75f6166aa5c017
Reviewed-on: https://chromium-review.googlesource.com/233770
Trybot-Ready: Paul Stewart <pstew@chromium.org>
Reviewed-by: Peter Qiu <zqiu@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>