shill: Ethernet: Do not call Stop() in destructor
In the common case where an Device destructor is called
as the device is being removed by Manager::DeregisterService(),
the device is already explicitly stopped.  In the exceptional
case when the Device destructor is called from the Manager
destructor (really the destructor for the vector of DeviceRefPtr
objects), it is unsafe to call Manager::DeregisterService() as
this will traverse a vector that has already been destroyed.

Instead, call the Stop() method explicitly on all devices in
Manager::Stop().

BUG=chromium:440767
TEST=Unit tests

Change-Id: Icc66f3793ccca932ad996b9b280eb89d975e4ae0
Reviewed-on: https://chromium-review.googlesource.com/234407
Reviewed-by: Peter Qiu <zqiu@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>