shill: allow configuring wifi services with kSecurityClassProperty
We want Chrome to use kSecurityClassProperty, instead of
kSecurityProperty.  To do so cleanly, Chrome needs to be able to
configure WiFi networks using the kSecurityClassProperty. So update
shill to support that.

While there: fix a buglet in WiFiService::SetPassphrase. (It was
logging the pointer to the error, rather than the error's message.)

BUG=chromium:441904
TEST=unit tests

Change-Id: I26957dbb7610381372044e3178a37b7b074764af
Reviewed-on: https://chromium-review.googlesource.com/235451
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>