shill: change wake to scan frequency member and property name
Add the "Seconds" unit to the wake to scan frequency member and property
name to explicitly state what unit this frequency property is given in.
Also, change the substring "Frequency" to "Period" to more accurately
reflect that this member/property refers to the length of time between each
wake to scan event.

BUG=chromium:448222
CQ-DEPEND=CL:240266
TEST=Compile shill.

Change-Id: I1b3dd8ca1a1172168047f25d912c44d6ab2b98aa
Reviewed-on: https://chromium-review.googlesource.com/240267
Tested-by: Samuel Tan <samueltan@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Samuel Tan <samueltan@chromium.org>