shill: add passive link monitoring support
PassiveLinkMonitor monitors the link by listening for ARP requests only.
It monitors the link by cycles, with each cycle being 25 seconds,
and is expected to receive at least 5 ARP requests in a cycle. When that
expectation is not met during a cycle, it will invoke the result callback
indicating link failure. The caller can specify the number of cycles to
passively monitor the link for. If that number of cycle is passed without
any failures, the result callback will be invoked to indicating success.
The PassiveLinkMonitor will automatically stop before the result callback
is invoked in either failure or success case.

The current link monitor implementation monitors the link by actively
sending ARP requests and waiting for ARP responses, which will be
refactor to a new class ActiveLinkMonitor in the upcoming CL.

The new LinkMonitor will then leverage both PassiveLinkMonitor and
ActiveLinkMonitor to monitor the link, with only one being active
at a time. This will reduce the number of ARP requests that the device
will send for link monitoring.

BUG=chromium:422159
TEST=USE="asan clang" FEATURES=test emerge-$BOARD shill

Change-Id: I3bfcaf4f49ae103764b2f79a609f669cfad89af8
Reviewed-on: https://chromium-review.googlesource.com/241713
Tested-by: Zeping Qiu <zqiu@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Zeping Qiu <zqiu@chromium.org>