shill: use map instead of vector for nested netlink attributes
Change underlying implementation of NetlinkNestedAttribute so that it stores
NestedData in maps instead of vectors. This simplifies the adding of nested
members in the constructor of nested attributes by making the order of insertion
of the nested data irrelevant, and removing the need to manually align the
NestedData in the underlying data structure so that their indices match the
attributes that they represent.

BUG=chromium:452732
TEST=Compile shill and run unit tests using:

  P2_TEST_FILTER="shill::*" FEATURES="test" USE="wake_on_wifi \
    clang asan" emerge-samus shill

  P2_TEST_FILTER="shill::*" FEATURES="test" USE="-wake_on_wifi \
    clang asan" emerge-samus shill

Change-Id: I1c378aecb19048804d8c4b5c3054e83db16bda92
Reviewed-on: https://chromium-review.googlesource.com/243780
Reviewed-by: Paul Stewart <pstew@chromium.org>
Reviewed-by: Samuel Tan <samueltan@chromium.org>
Tested-by: Samuel Tan <samueltan@chromium.org>
Commit-Queue: Samuel Tan <samueltan@chromium.org>