shill: WiFiService: Log only interesting endpoint updates
Only log endpoint changes if the BSSID, signal strength or
frequency will change as a result of the update.  This is of most
benefit when scan results provide multiple endpoints for a single
service.  With this change only frequency or signal strength
changes in the currently selected endpoint or a change to a
different endpoint will trigger this message.

BUG=chromium:458225
TEST=Unit tests for regression

Change-Id: Ibf894d5653e638e13555b7fd759eb26bc884226c
Reviewed-on: https://chromium-review.googlesource.com/249451
Reviewed-by: Rebecca Silberstein <silberst@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>