shill: disable DeviceTest.PrependIPv6DNSServers
The shill_unittest binary as produced by the next_gcc compiler currently
fails to complete this test on the release buildbots.  This is blocking
the release of the next_gcc toolchain, so the test will be disabled
temporarily until a solution found for the builders.

BUG=chromium:462647
TEST=FEATURES=test emerge-samus shill

Change-Id: I4ae2d3ce506e62e9f26c53670edf6b74f4759277
Reviewed-on: https://chromium-review.googlesource.com/254610
Commit-Queue: Garret Kelly <gdk@chromium.org>
Trybot-Ready: Garret Kelly <gdk@chromium.org>
Tested-by: Garret Kelly <gdk@chromium.org>
Reviewed-by: Luis Lozano <llozano@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>