shill: remove progressive scan variation support
We're not tracking the data from the progressive scan field
trial, so we should just remove it.

(Doing so now is motivated by the fact that we're removing the
manual control over progressive scan in crosh. So there's no
longer an escape hatch for a device that encounters problems
due to the field trial.)

BUG=chromium:250945
TEST=FEATURES=test emerge-peppy shill

Change-Id: I5e82e75e0f95e33a89c97abf9b33eab6157de08e
Reviewed-on: https://chromium-review.googlesource.com/272219
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Zeping Qiu <zqiu@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>