shill: dhcp: Tolerate early dhcpcd exit
When the shill process is killed (e.g., during shutdown) a
terminate signal is sent to the shill process group, which
includes the dhcpcd process.  As the two processes race to
exit, it is common that the dhcpcd process is already gone
by the time shill attempts to release its lease and kill
the dhcpcd subprocess.  Neither of these errors are worthy
of concern, since it is known ahead of time that we are in
the process of stopping dhcpcd when these errors occur.

BUG=chromium:498908
TEST=Unit tests (for regression)
TEST="restart shill; grep dhcp /var/log/messages"
TEST=Should contain no messages of the form:
TEST="dhcpcd daemon appears to have already exited" or
TEST="No such process"

Change-Id: I48d2e63c9793eb7f635d2973d931e94c5e8227a2
Reviewed-on: https://chromium-review.googlesource.com/277136
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Zeping Qiu <zqiu@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>