shill: DHCPConfig: Ignore 576 as an MTU value
The minimum IPv4 MTU value, 576, is also used as a failsafe
"I don't know what the MTU is" value.  Unfortunately this is
used often enough that it is no longer of practical value as
a parameter to be honored by the DHCP client.  This CL adds
logic to ignore this specific value in DHCP negotiation.  It
also adds a metric so we can monitor what values appear in the
wild.

As a consequence of this change, DHCP-based networks that
actually have an MTU of 576 will likely not work correctly.
For the network types that this change affects in shill,
namely WiFi, LTE and Ethernet, this is unlikely to be the case.
Nonetheless, it should be noted that such unfortunate networks
will be affected.

This method is implemented in parallel to the minimum MTU
command-line option in order to separate the concerns -- the
minimum MTU parameter specifies a range of MTU values, while
this change is meant to highlight a specific value that should
be disregarded.

BUG=chromium:498358
TEST=Unit test

Change-Id: Ice7310b2ad7dc0092a80d0953757109e956fa4cd
Reviewed-on: https://chromium-review.googlesource.com/277455
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>