shill: DHCPCDProxy: Reduce severity of rebind failures
The Rebind operation often happens at resume-from-suspend while
a number of other activities (including link loss) are occurring.
As such, it's possible for dhcpcd to have already exited by the
time shill attempts a rebind, due to the disconnect event not
yet being acted on by shill.  In such a conditon it is safe to
ignore a failure in Rebind since another event (either the exit
of the child process or loss of link on the interface) will also
provide the same information.

BUG=chromium:496883
TEST=Unit tests

Change-Id: I99b04e151fb22a84056b4e2e9d8905b10823bfd2
Reviewed-on: https://chromium-review.googlesource.com/281625
Reviewed-by: Zeping Qiu <zqiu@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>