shill: change criteria for recording dark resume events
Log dark resume events for purposes of throttling when the
current dark resume is caused by wake on SSID or wake on
disconnect, since we want to guard against repeated dark resume
wakes caused by these triggers.

We previously logged only dark resumes that we entered in a
disconnected state. However, this was an issue when shill
entered dark resume because of a disconnect---shill would only
"realize" that it is disconnected shortly after entering dark resume.
This new criteria fixes this issue by relying on the wake reason,
not connectivity status, to decide whether or not to log the
current dark resume.

Also, improve wake on WiFi throttling unit tests by mocking the
time values logged in |dark_resume_history_|.

BUG=chromium:505072
TEST=Compile shill and run unit tests.

Change-Id: I9b99fcea3d1245e5a3db1e6510e24126a48d8d8b
Reviewed-on: https://chromium-review.googlesource.com/282522
Tested-by: Samuel Tan <samueltan@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Samuel Tan <samueltan@chromium.org>