shill: track DisconnectReason from supplicant
Supplicant reports a disconnect reason when handling a disconnect event.  This
reason code can be used to troubleshoot and diagnose connectivity issues.  Each
DisconnectReason property update is now processed by shill.  Shill logs and
stores the reason code for the duration of the disconnected state of a service.
When the current or pending service moves from the disconnected state, the value
is cleared.  The default value for the disconnect reason is an invalid reason
code to distinguish between explicitly set and default reason codes.

BUG=chromium:508225
TEST=Unit tests
TEST=manually tested on peppy

Change-Id: Ia590bd17e6086633ba29448d39b999379f8270f7
Reviewed-on: https://chromium-review.googlesource.com/284624
Trybot-Ready: Rebecca Silberstein <silberst@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: Rebecca Silberstein <silberst@chromium.org>
Tested-by: Rebecca Silberstein <silberst@chromium.org>