regulatory-domain: Depend on network-services
Move the loading of the cfg80211 module over to the
network-services upstart task, so that tasks like
regulatory-domain don't have to directly depend on
shill.  While here, also make this a template that
is able to disable loading cfg80211 on hosts that
do not have WiFi enabled.

CQ-DEPEND=CL:284174,CL:284740
BUG=brillo:1239
TEST=emerge-$BOARD shill && cros deploy $DUT shill && ssh $DUT reboot
TEST=ssh $DUT lsmod  # <- make sure cfg80211 is listed

Change-Id: I2778eb15399625860b7e04f67f9681602f47c866
Reviewed-on: https://chromium-review.googlesource.com/284701
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>