shill: ByteString: Get rid of RemovePrefix()
This method only had one remaining user, and it was not a
compelling use case.  Compared with the technical cost of
maintaining this non-intuitive begin_ pointer, maintaining
RemovePrefix() was not worth it.  This CL removes the
implementation and fixes the last user to not require it.

Also, while here, fix buffer read overrun issue in the
related code.

BUG=None
TEST=Unit tests + manual connect to WiFi network

Change-Id: I4efcebeb49ceb370e69b1b0e8f6b9064995b00da
Reviewed-on: https://chromium-review.googlesource.com/289350
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Queue: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>