shill: abstraction for remote service watcher
This interface will monitor on the owner (unique connection name)
of a remote service, the callback will be invoked when this owner
exits.

This is needed by DeviceClaimer for monitoring on the remote service
that claims interfaces from shill.  The unique connection name of
the claimer will be obtained through the raw dbus message provided
in the method handler.

BUG=chromium:517680
TEST=USE="chromeos_dbus asan clang" FEATURES=test emerge-$BOARD shill

Change-Id: I3d0c1746340cf45f9536fe69ede3afbd6d89234f
Reviewed-on: https://chromium-review.googlesource.com/291455
Commit-Ready: Zeping Qiu <zqiu@chromium.org>
Tested-by: Zeping Qiu <zqiu@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>