shill: JsonStore: add Close() method
For plug compatibility with KeyFileStore, we need to provide
Close().

As for what Close() should actually do... It's clear that Close()
needs to Flush() the in-memory data to disk. But beyond that,
things are less clear.

Given the comment above Open(), in key_file_store.h, it seems
that calls to setters or getters should fail on a closed store.
It's not quite clear to me why KeyFileStore chose to provide
those semantics. And doing the same would require extra work
in JsonStore (keeping track of the open/close state, and checking
it before any Get/Set operations).

For now, I'm implementing Close() as just a Flush(). We can
revisit this, if that causes problems.

While there: fix style issue with some API comments.

BUG=b:22958135
TEST=unit tests

Change-Id: I3df033e7cac85bdf3d6e2cd838eb0cc41672404d
Reviewed-on: https://chromium-review.googlesource.com/293593
Trybot-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>