shill: DeviceInfo: Delay test for Tunnel devices
It's been observed that tunnel devices can be mis-detected by
shill because the tun_flags sysfs meta-file is created after
the RTNL event arrives to indicate the presence of the device.
To solve this, move detection of tunnel devices to the
DelayedDeviceCreationTask so the meta-file is created by the
time a decision is made about the device type.

BUG=chromium:524164
TEST=Unit tests
TEST=Run "openvpn --mktun --dev tun0" while shill is running.
TEST=Wait 5 seconds.  Do an "ifconfig -a".  "tun0" should not be there.

Change-Id: Ieb148cbb4bd6a7aeda0bb5402579e53a14dcadfd