shill: ManagerTest: isolate references to GLib
Tests that need to persist/restore profile data will
need to use GLib, if we're building with KeyFileStore.
But having the instantiation of GLib sprinkled into
each of those test cases makes it hard to compile out
GLib in cases where we don't need it.

Fix this, by moving instantiation of GLib to the
ManagerTest ctor. While we're at it, make this
instantiation conditional on whether or not we're
using KeyFileStore.

Note that the test fixture now has two GLib references.
One GLib reference, inherited from PropertyStoreTest,
points to a MockGLib. The new reference will point
either to a regular (non-mock) GLib, or to null.

BUG=b:23386647
TEST=unit tests

Change-Id: I525925a2919525c4e47dd77c9f2c23b5985e31f7
Reviewed-on: https://chromium-review.googlesource.com/295546
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>
Commit-Queue: mukesh agrawal <quiche@chromium.org>