shill: start a loop to handle shill's async termination tasks
Currently, chromeos::Daemon will start a secondary loop on shutdown
to process asynchronous tasks. It only waits for the tasks that are
posted on the loop, but not pending IO operations, such as async
D-Bus method calls.  This causes the process to exit before
termination tasks are completed, such as syncing profiles to disk.

So start another loop for handling shill's termination tasks, it
will break out of the loop when all termination tasks are completed.

BUG=chromium:517680
TEST=USE="chromeos_dbus cellular asan clang" FEATURES=test emerge-$BOARD shill
TEST=Run network_3GSmokeTest
CQ-DEPEND=CL:295772

Change-Id: I48d83d7bb3a14b4f3aee838742b697e6f5a44507
Reviewed-on: https://chromium-review.googlesource.com/295538
Commit-Ready: Zeping Qiu <zqiu@chromium.org>
Tested-by: Zeping Qiu <zqiu@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>