shill: KeyFileStoreTest: dynamically allocate store
We want to remove the set_path() method from StoreInterface,
and have StoreInterface clients pass the path to the ctor
instead. But KeyFileStore is an inline member of
KeyFileStoreTest. So, if we add a parameter to the
KeyFileStore ctor, then the test class will need to
supply that parameter in the initializer list.

There's no good way to supply that parameter at the
time KeyFileStoreTest is being constructed, since the
parameter requires state that doesn't exist at construction
time. (Specifically: we need to call CreateUniqueTempDir()
on the ScopedTempDir. But we can't do that in the initializer
list.)

We could try to kludge around this by, e.g., adding an
AutomaticScopedTempDir class, which calls its own
CreateUniqueTempDir() method in its own ctor. But, given
that the temp dir creation currently happens in SetUp(),
rather than the KeyFileStoreTest ctor, that seems like a
philosophical change.

Instead, we simply allocate the KeyFileStore in
KeyFileStoreTest::SetUp().

This change doesn't really stand on its own... But it's
done separately from the CL for removing set_path(), simply
because of the size of this change. Breaking out this
(largely mechanical) change into a separate CL should make
it easier to understand the set_path() CL.

BUG=chromium:517570
TEST="USE='asan clang' FEATURES=test emerge-samus shill"

Change-Id: I7464026923ef709004fe476865955a9d3abd475c
Reviewed-on: https://chromium-review.googlesource.com/295857
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>