shill: JsonStoreTest: dynamically allocate store
As with KeyFileStoreTest, we won't be able to
statically allocate the store, once the JsonStore
constructor requires a path argument.

Pave the way for that shining future, by allocating
the store dynamically, and updating all code that
references the store.

While the CL doesn't make sense in isolation, it's
done separately to keep the add-arg-to-ctor CL small
and clean.

BUG=chromium:517570
TEST="USE='asan clang json_store' FEATURES=test emerge-samus shill"

Change-Id: I45b4504a5d50f44d19182a207a5afccb0e819229
Reviewed-on: https://chromium-review.googlesource.com/295858
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>