shill: KeyFileStoreTest: remove unnecessary path change
It's unclear what we accomplish by changing the path
during fixture TearDown. While the comment indicates
that this is a way of avoiding saving changes, that
doesn't quite make sense, for a couple reasons:

1) Why should we care if the store is saved or not?
2) The only KeyFileStore code that should run after
   the fixture TearDown() is KeyFileStore dtor. But
   that doesn't pay any attention to |path_|. Nor
   does the dtor call Close() or Flush().

In fact, v1 of the patch that landed this code had
an explicit store_.Close(), instead of any attempt
to avoid saving the store. AFAICT, simply removing
that Close() call is all we need to do, to avoid
writing the store to disk.

BUG=chromium:517570
TEST="USE='asan clang' FEATURES=test emerge-samus shill"

Change-Id: Ib64e339635cb5d898a0ddc9e7abbed16197f542a
Reviewed-on: https://chromium-review.googlesource.com/295928
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>