shill: KeyFileStoreTest: don't use empty path
Now that KeyFileStore takes path as an argument to the
ctor, and validates that the path is non-empty, we'd
like to get rid of the empty-path checks in KeyFileStore.

Update the test code to no longer require the empty
path checks. Instead, force Flush() to fail in a more
fundamental way.

BUG=None
TEST="USE='asan clang' FEATURES=test emerge-samus shill"

Change-Id: I58c4187243c86b760dce555f8a7763648578e6b0
Reviewed-on: https://chromium-review.googlesource.com/295927
Commit-Ready: mukesh agrawal <quiche@chromium.org>
Tested-by: mukesh agrawal <quiche@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>