shill: mock out NetlinkManager init functions
Currently, NetlinkManager's init functions (Init and Start) is not
being mocked out in unittest, and will attempt to create socket
during test. For non-x86 boards, the socket creation would fail.
I think this is because the host is a x86 machine.  With the new
message loop, an invalid file descriptor is not allowed.

Fix it by mocking out NetlinkManager's Init and Start function for
unittests.  Also add the missing include for base::Closure in the
ChromeosDaemon's header

BUG=chromium:517680
TEST=FEATURES=test emerge-daisy shill
CQ-DEPEND=CL:295772

Change-Id: I8fc86ff5dfbf8c98a940c38b8e5a38d288a914b8
Reviewed-on: https://chromium-review.googlesource.com/295932
Commit-Ready: Zeping Qiu <zqiu@chromium.org>
Tested-by: Zeping Qiu <zqiu@chromium.org>
Reviewed-by: Paul Stewart <pstew@chromium.org>