shill: use 'json' extension for JsonStore profiles
To avoid any confusion between files using different formats,
we want KeyFileStore profiles and JsonStore profiles to have
different names. So add '.json' to the file name when using
JsonStore as the backend.

Note: initially, I wanted to isolate the ENABLE_JSON_STORE
dependent code to just StoreFactory. But I've come to change
my mind on that.

The reason for my change of mind is that {Default}ProfileTest
checks the actual path that will be used when writing the file.
And that seems like an appropriate thing to test. But, to do
that, {Default}ProfileTest needs to know when backend we're
using.

So we need ENABLE_JSON_STORE checks in {Default}ProfileTest.
And once we've done that, it seems silly, and unnatural, to
hide this information from Profile.

Along the way:
- Move path computation from InitStorage() and RemoveStorage(),
  into the {Default}Profile constructor. This makes the code
  slightly simpler (removing a virtual method) and shorter.
- Provide a method for tests to query the path that Profile
  would use, so that we don't duplicate path computation logic.
- Remove |storage_path_| field, since it was only needed to
  compute the path of the persistent file.

While there:
- Rename |store_path| argument to DefaultProfile ctor. It's
  unclear whether such a "path" should refer to a directory
  or a file.
- Rename |user_storage_directory| argument to Profile ctor.
  This name makes no sense when the Profile ctor is being
  called from the DefaultProfile ctor.

Bug: 23561491
TEST="USE=json_store FEATURES=test emerge-samus shill"

Change-Id: I9b7e9b04dc80c89c20ea43158ed293b8a50fadbf