shill: use empty string for null profile RPC ID
During Chrome OS D-Bus migration, profile RPC ID was incorrectly assumed
as an object type instead of a string. This causes "/" to be used as
null RPC ID instead of empty string, since an empty string is an invalid
object path.

Fix this by going back to use an empty string as null RPC ID for profiles
and update the documentation to reflect the actual type.

BUG=chromium:527902
TEST=unittests
TEST=Manual test: verify Preferred Networkds doesn't include all networks

Change-Id: I1ef691f24f7e34eda4d3e929c90ed0bf5db6adf4
Reviewed-on: https://chromium-review.googlesource.com/297188
Commit-Ready: Zeping Qiu <zqiu@chromium.org>
Tested-by: Zeping Qiu <zqiu@chromium.org>
Reviewed-by: Zeping Qiu <zqiu@chromium.org>