shill: Add libshill and libshill-client target
Add the libshill and libshill-client targets to the shill Android.mk file.
Also, add the supplicant-proxies and dhcpcd-proxies targets that
are required by the libshill target.

The following source files used to build the libshill target in shill.gyp
are not included in the Android.mk file:
- key_file_store.cc (using json_store.cc instead)
- crypto_des_cbc.cc, crypto_provider.cc (not using key_file_store.cc)
- source files required for vpn (not using VPN)
- source files required for cellular (not using cellular)
- source files required for wimax (not using wimax)
- dbus_bindings/upstart.dbus-xml,
  dbus/chromeos_upstart_proxy.cc (not using upstart)
- dbus/chromeos_power_manager_proxy.cc (not using power manager)
- dbus/chromeos_permission_broker_proxy.cc (not using permission broker)

The following source files are additionally used to build libshill in
Android.mk, but are not included shill.gyp:
- dbus/chromeos_firewalld_proxy.cc (use firewalld instead of permission
  broker)
- power_manager_proxy_stub.cc (not using power manager)
- upstart/upstart_proxy_stub.cc (not using upstart)

BUG: 22956441
Change-Id: Ia1388654fbf15f04e72ba704e4f514682389854e
TEST: cd into shill directory, run 'mm' and make sure that all builds succeed.