Build shill from the new directory path
The shill ebuild recreates the old directory structure placing
shill as a subdirectory of platform2 in the build environment.
This structure doesn't match the current location of the code and
breaks incremental builds creating symlinks to shill in
platform2/ and even in the new shill directory.

This patch modifies the .gyp file to reference platform2 sources via
the platform2_root gyp variable and builds shill from the new
path. This fixes incremental builds.

For this change to take effect in Chrome OS, the ebuild file is updated
in a separated CL.

Bug: 23587594
TEST=emerge-$BOARD shill; cros_workon_make shill

Change-Id: I3ccacd97134c792dc5f372e00712f3b885892335