shill: nl80211: Accept U8 types for Nl80211AttributeRegInitiator
Solve an issue with message parsing.  The RegInitiator attribute
as observed from packets is at least sometimes 1 byte long, not 4.
From just this observation it could be assumed that this value
should be interpreted as a U8.  From reading through the "iw"
source code (in event.c, print_event() has multiple handling cases
for NL80211_ATTR_REG_INITIATOR) however, it's apparent that this
value can be interpreted as either a U8 or U32 depending on
context.

This CL overrides the InitFromValue() method so that a 1-byte
value is not considered in error.

BUG: 24007462
Change-Id: I91c92252e15f5ce11d00aaf3d03e6864772903e3
TEST: Restart shill with WiFi, look for messages from b/24007462