shill: ProcessManager: add UpdateExitCallback
We want to migrate CryptoUtilProxy and OpenVPNDriver
from ProcessKiller to ProcessManager. There a hitch,
though.

Both classes use ProcessKiller with an exit callback.
And both classes want child exit to be handled differently
in this case, than before they've sent the kill signal.

To meet this need, we add UpdateExitCallback to
ProcessManager. With this new method, CryptoUtilProxy
and OpenVPNDriver will be able to retain their existing
behavior as follows:

	 process_manager_->UpdateExitCallback(
		pid, base::Bind(HandleChildDeath));
	 process_manager_->StopProcess(pid);

Bug: None
BUG=chromium:533594
TEST=unit tests

Change-Id: I3452c30088917b79311a3b3356bd7135a961be26