shill: ProcessManager: add StartProcessInMinjailWithPipes()
To migrate CryptoUtilProxy to ProcessManager, we need
ProcessManager to provide a method for forking a child
with pipes back to us.

Make it so.

While there: rename the |from_here| paramater to
|spawn_source|. I found it confusing to pass |from_here|
through to another function. (The name suggests that
we're passing the current location, rather than the
original location.)

Bug: None
BUG=chromium:533598
TEST=unit tests

Change-Id: Id89371d8700b693afed05be61e6913a1d0a496d4