shill: OpenVPNDriver: reduce use of GLib
We want to compile this code on boards that don't have
glib (the library). As part of that effort, fork
children using ProcessManager, instead of GLib
(the object).

While there: refactor ParseLSBRelease() and
InitEnvironment(). We haven't found any other use
for the LSB info, so it's simpler to just merge these
two methods.

Bug: None
BUG=chromium:528051
TEST=unit tests

Change-Id: Icc5402dbae56b2c58fb76156db93ba6fd3d7a1e9