shill: Profile: change ctor arg type from string to FilePath
The |user_storage_directory| argument to the Profile ctor
is stored inside the Profile object as a FilePath, and the
DefaultProfile ctor takes its path argument as a FilePath.

Both of these facts argue for |user_storage_directory| to
be passed in as a FilePath, rather than a string.

Make it so.

Bug: None
TEST=unit tests

Change-Id: I6094aeb44de96ccd0312c5311bdd195ee64cf510