libshill-net: add a wrapper class for WiFi driver HAL
On Brillo devices, it is necessary to manuallay initialize the
WiFi driver and configure the WiFi interface via the driver HAL
APIs.

Include this wrapper  as a part of libshill-net shared library
to allow other daemons such as apmanager to have access to those
APIs.

WiFi driver is reinitialized everytime we reconfigure the interface
mode.  This is for making sure the driver is initialized before
we attempt to configure the interface.  The driver will not perform
any initialization if it is already initialized.  So there won't be
any performance penalty in this case.

Bug: 24294634
TEST=Manual test using apmanager on a dragon board

Change-Id: I2a145bc3aad2c399599f50944b36ca0e33ede119