shill-test-proxy: Create Test Proxy daemon
Initial CL to create a skeletal framework for testing Shill via external
RPC calls on Brillo platforms.

There are 2 threads running on the daemon:
1. Main chromeos-dbus thread which runs the message loop and is used for
sending dbus commands to Shill.
2. RPC Server thread which blocks indefinitely waiting for external RPC
requests and dispatches tasks to run in the main chromeos-dbus thread.

The current implementation just uses the shill/setup_wifi code to
connect/disconnect to a given wifi ssid when we receive the ConnectWifi/
DisconnectWifi RPC requests.

BUG: 24335496
TEST: Compilation of the proxy is succcessful.

Change-Id: I964aaae88323a0b6963497152c6dd09d4015312d