shill: ignore devices with locally administered MAC addresses
Currently, the Brillo dragonboard incorrectly reports that its
p2p interface is a managed interface. Add a checks whether or
not the MAC address of the interface is locally or globally
administered to determined whether to create a WiFi device for
it.

This fix will be removed once the drivers for the dragonboard
correctly report the interface type for p2p.

BUG: 24372451
TEST: Run shill on dragonboard, read shill logs to ensure that WiFI
      device is not created for p2p0.

Change-Id: I1dfc07c3d51afcdb81e9e3a7b482668c222e3235