Support DNS server proxy on Android
There is no support for /etc/resolv.conf on Android, so a DNS
server proxy is needed to handle/redirect client DNS requests.
dnsmasq will be used to setup as a DNS server proxy, and it will
run under the "system" user.

Also add a special handling for the case when DHCP doesn't provide
DNS server addresses. We will default to use Google's DNS server
addresses in this case.

Not sure why we never had a problem with this scenario on Chrome OS,
so make this handling Android specific for now.

TODO: update connection_unittest.cc for Android.

Bug: 24543278
TEST=Manual test on dragonboard (ping to well known hosts)

Change-Id: I4573546a9987b1148a9c0e6157155a232be4f168