shill-test-proxy: Make test-proxy single threaded.
The test proxy should process RPC requests sequentially and
synchronously. So, there is no need for the proxy to be multi-threaded
and add the overhead of synchronizing the 2 threads. The RPC server
holds the pointer to a single instance of the ShillWifiClient and uses it
to execute the dbus commands needed to process an RPC request.

While there:
1. Include the correct chromeos_variants.h header instead of
chromeos_any.h in the ProxyShillWifiClient.
2. Change all instance of |timeout_in_seconds| to |timeout_seconds| in
function params to be consistent.

Bug: 24335496
Change-Id: Ice574f946e127ab670a0dee575a4467cf9ca8b51
TEST: mmm system/connectivity/shill/test-rpc-proxy