Do not export WiFiDrvierHal APIs
The original intent for exporting these APIs is allow other
daemons (e.g. apmanager) to setup WiFi interface in different
modes.  This is not needed anymore since it is already being
handle through NL80211 commands (used by wpa_supplicant and
hostapd).

We only expect these APIs to be used by shill for the initial
WiFi driver setup. So move it to the shill's internal library
instead of being in the shared library.

Bug: None
TEST=Verify WiFi still works on dragonboard

Change-Id: I94f8531b47b9d616478b3117f68485c078fb134d