dbus: fix duplicate serivce appeared callback for PowerManager
Currently, we registered two service monitoring callbacks when
initializing the PowerManagerProxy.  First one is an one-time callback
which is invoked when service becomes available
(WaitForServiceToBeAvailableCallback), second one
(NameOwnerChangedCallback) is invoked whenever service owner changes
(monitored continuously).

This setup have an undesired side effects, where if the service comes
up after the callbacks are registered, both callbacks will be invoked.
This causes the |service_appeared_callback_| to be invoked twice, and
causes PowerManager to generate FATAL error on the second invocation.
If the service is already up when the callbacks are registered, only
WaitForServiceToBeAvailableCallback will be invoked, since there is
no changes to the service owner.

The fix is to only register WaitForServiceToBeAvailableCallback during
proxy initialization, this callback is meant to be used for initial
service detection. Then register the NameOwnerChangedCallback
to continuously monitor the service owner changes when the service becomes
available.

Bug: None
BUG=chromium:541647
TEST="start powerd" after shill is started, verify shill doesn't crash

Change-Id: I79b6aa237893d9b92363d1fc6441d933e57296ca