shill: add SHILL_USE_WIFI flag
Add SHILL_USE_WIFI flag, which causes shill to compile
with WiFi enabled iff this flag is set to true.

While there, change all shill_use_* flags to upper case,
and check for the value "true" instead of "yes". Since
there are currently no users of the shill_use_* flags,
this change is safe to make.

BUG: 24749702
TEST: Built for a target that sets neither SHILL_USE_WIFI nor
      SHILL_USE_WIRED_8021X and ensured that it excluded the
      EAP and supplicant source files and built successfully.
      and ensure that shill builds and runs without WiFi errors.
TEST: Built for target that sets SHILL_USE_WIFI to "true" and
      ensure that EAP and supplicant source files are included
      and that shill builds successfully.

Change-Id: Icf85c7862daf5c6ba8d8209d44d66459156350e4