shill: re-enable/disable compiler warnings used in CrOS on Android
Currently, shill builds with different compiler warnings
enabled/disabled on CrOS and Android. Disable and re-enable
some of these compiler warnings when building for Android so
that shill will build with the same warnings disabled on
both CrOS and Android.

While there, fix indentation in Android.mk and remove
'-Wno-sign-promo' since it is unnecessary.

BUG: 24985776
Change-Id: I25e61e7cc5706877585907d37959a7ed02036ce6
TEST: shill builds successfully on CrOS and Android.